refactor post redis
